+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/fixture/MockSynchronizer.kt
@@ -33,6 +33,7 @@ import io.ktor.client.HttpClientConfig
 import io.ktor.client.engine.HttpClientEngineConfig
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
+import kotlin.time.Duration
 
 /**
  * Mocked Synchronizer that can be used instead of the production SdkSynchronizer e.g. for tests.
@@ -279,6 +280,19 @@ internal class MockSynchronizer(
         set(value) {}
 
     override suspend fun getFastestServers(servers: List<LightWalletEndpoint>): Flow<FastestServersResult> {
+        error("Intentionally not implemented in ${MockSynchronizer::class.simpleName} implementation.")
+    }
+
+    override suspend fun evaluateServerSwitch(
+        current: LightWalletEndpoint,
+        candidates: List<LightWalletEndpoint>,
+        fetchThreshold: Duration,
+        blocksToFetch: Int
+    ): LightWalletEndpoint? {
+        error("Intentionally not implemented in ${MockSynchronizer::class.simpleName} implementation.")
+    }
+
+    override suspend fun confirmServerSwitch(endpoint: LightWalletEndpoint) {
         error("Intentionally not implemented in ${MockSynchronizer::class.simpleName} implementation.")
     }